+++ core/api_container/service_network/service_network.go
@@ -220,6 +220,7 @@ func (network *ServiceNetwork) AddServiceInPartition(
 	defer func() {
 		if !shouldLeaveServiceInTopology {
 			network.topology.RemoveService(serviceId)
+			network.freeIpAddrTracker.ReleaseIpAddr(serviceIp)
 			delete(network.serviceIps, serviceId)
 
 			// NOTE: As of 2020-12-31, we don't actually have to undo the iptables modifications that we made to all
@@ -386,6 +387,10 @@ func (network *ServiceNetwork) RemoveService(
 	if !found {
 		return stacktrace.NewError("Unknown service '%v'", serviceId)
 	}
+	serviceIp, found := network.serviceIps[serviceId]
+	if !found {
+		return stacktrace.NewError("No IP found for service '%v'", serviceId)
+	}
 
 	// TODO Parallelize the shutdown of the service container and the sidecar container
 	// Make a best-effort attempt to stop the service container
@@ -394,7 +399,7 @@ func (network *ServiceNetwork) RemoveService(
 		return stacktrace.Propagate(err, "An error occurred stopping the container with ID %v", serviceContainerId)
 	}
 	network.topology.RemoveService(serviceId)
-	// TODO release the IP that the service got
+	network.freeIpAddrTracker.ReleaseIpAddr(serviceIp)
 	delete(network.serviceContainerIds, serviceId)
 	delete(network.serviceIps, serviceId)
 	logrus.Debugf("Successfully removed service with container ID %v", serviceContainerId)
@@ -414,6 +419,7 @@ func (network *ServiceNetwork) RemoveService(
 
 		}
 		sidecarContainerId := sidecarContainerInfo.containerId
+		sidecarIp := sidecarContainerInfo.ipAddr
 
 		// TODO Parallelize the shutdown of the sidecar container with the service container
 		// Try to stop the sidecar container too
@@ -422,7 +428,7 @@ func (network *ServiceNetwork) RemoveService(
 		if err := network.dockerManager.KillContainer(context, sidecarContainerId); err != nil {
 			return stacktrace.Propagate(err, "An error occurred stopping the sidecar container with ID %v", sidecarContainerId)
 		}
-		// TODO release the IP that the service received
+		network.freeIpAddrTracker.ReleaseIpAddr(sidecarIp)
 		delete(network.sidecarContainerInfo, serviceId)
 		logrus.Debugf("Successfully removed sidecar container with container ID %v", sidecarContainerId)
 	}